+++ bevy_ios_iap/src/native.rs
@@ -132,18 +132,18 @@ mod ffi {
             original_id: u64,
             json_representation: String,
             product_type: IosIapProductType,
-            reason: IosIapTransactionReason,
             environment: IosIapEnvironment,
-            storefront: IosIapStorefront,
         ) -> IosIapTransaction;
+        #[swift_bridge(associated_to = IosIapTransaction)]
+        fn add_storefront(t: &mut IosIapTransaction, store: IosIapStorefront);
+        #[swift_bridge(associated_to = IosIapTransaction)]
+        fn add_reason(t: &mut IosIapTransaction, reason: IosIapTransactionReason);
         #[swift_bridge(associated_to = IosIapTransaction)]
         fn add_revocation(t: &mut IosIapTransaction, date: u64);
         #[swift_bridge(associated_to = IosIapTransaction)]
         fn add_expiration(t: &mut IosIapTransaction, date: u64);
         #[swift_bridge(associated_to = IosIapTransaction)]
         fn add_currency(t: &mut IosIapTransaction, currency: IosIapCurrency);
-        #[swift_bridge(associated_to = IosIapTransaction)]
-        fn add_currency_code(t: &mut IosIapTransaction, code: String);
         #[swift_bridge(associated_to = IosIapTransaction)]
         fn revocation_reason(t: &mut IosIapTransaction, reason: IosIapRevocationReason);
         #[swift_bridge(associated_to = IosIapTransaction)]